refactor: 게시글 삭제 시 유저 일치 여부 확인

+++ src/post/post-diet.service.ts
@@ -1,8 +1,8 @@
 import {
-  ForbiddenException,
   Inject,
   Injectable,
   NotFoundException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -32,11 +32,11 @@ export class PostDietService {
         createPostDto,
       )}`,
     );
-    console.log(createPostDto.userEmail);
+
     const user = await this.usersRepository.findOne({
       where: { email: createPostDto.userEmail },
     });
-    console.log(user);
+
     if (!user) {
       throw new NotFoundException(
         `이메일이 ${createPostDto.userEmail}인 사용자를 찾을 수 없습니다. `,
@@ -96,12 +96,12 @@ export class PostDietService {
       throw new NotFoundException(`PostID ${postId}를 찾을 수 없습니다.`);
     }
 
-    if (foundPost.user.email != updatePostDto.email) {
-      throw new ForbiddenException(`작성자가 아닙니다!`);
-    }
-
     // 메뉴와 이메일 분리
     const { menues, email, ...data } = updatePostDto;
+
+    if (foundPost.user.email != email) {
+      throw new UnauthorizedException(`게시글을 수정할 권한이 없습니다.`);
+    }
 
     // 객체를 덮어씌우기
     Object.assign(foundPost, data);
@@ -134,14 +134,19 @@ export class PostDietService {
     return savedPost;
   }
 
-  async remove(id: number) {
+  async remove(id: number, userEmail: string) {
     this.logger.info(`diet post remove service called with Post ID: ${id}`);
     const foundPost = await this.postsRepository.findOne({
       where: { id },
+      relations: ['user'],
     });
 
     if (!foundPost) {
       throw new NotFoundException(`게시글 ID ${id}을 찾을 수 없습니다.`);
+    }
+
+    if (foundPost.user.email !== userEmail) {
+      throw new UnauthorizedException(`게시글을 수정할 권한이 없습니다.`);
     }
 
     await this.postsRepository.remove(foundPost);
